Add initial implementation of components and events, including Widget and Event classes

+++ src/events.cpp
@@ -1,22 +1,24 @@
 #include <iostream>
 #include <string>
 
-#include "../include/events.hpp"
+#include "events.hpp"
 
 // MouseEvent implementation
-MouseEvent::MouseEvent(int x, int y, int button) : _x(x), _y(y), _button(button) {}
+MouseEvent::MouseEvent(u_int16_t x, u_int16_t y, int8_t button)
+    : _x(x), _y(y), _button(button) {}
 
 void MouseEvent::Handle()
 {
         std::cout << "Mouse Event: (" << _x << ", " << _y << ") Button: " << _button << std::endl;
 }
 
-int MouseEvent::GetX() const { return _x; }
-int MouseEvent::GetY() const { return _y; }
-int MouseEvent::GetButton() const { return _button; }
+u_int16_t MouseEvent::GetX() const { return _x; }
+u_int16_t MouseEvent::GetY() const { return _y; }
+int8_t MouseEvent::GetButton() const { return _button; }
 
 // KeyboardEvent implementation
-KeyboardEvent::KeyboardEvent(char key) : _key(key) {}
+KeyboardEvent::KeyboardEvent(char key)
+    : _key(key) {}
 
 void KeyboardEvent::Handle()
 {
@@ -26,9 +28,20 @@ void KeyboardEvent::Handle()
 char KeyboardEvent::GetKey() const { return _key; }
 
 // EventManager implementation
-void EventManager::ProcessEvent(Event &event) { event.Handle(); }
-void EventManager::ProcessMouseEvent(MouseEvent &mouseEvent) { mouseEvent.Handle(); }
-void EventManager::ProcessKeyboardEvent(KeyboardEvent &keyboardEvent) { keyboardEvent.Handle(); }
+void EventManager::ProcessEvent(Event &event)
+{
+        event.Handle();
+}
+
+void EventManager::ProcessMouseEvent(MouseEvent &mouseEvent)
+{
+        mouseEvent.Handle();
+}
+
+void EventManager::ProcessKeyboardEvent(KeyboardEvent &keyboardEvent)
+{
+        keyboardEvent.Handle();
+}
 
 // Test
 int main()
